refacor : redux 폴더구조 변경
