--- client/src/redux/action-types/index.ts
@@ -1,33 +1,0 @@
-export enum ActionType {
-  GET_PRODUCTS_REQUEST = "GET_PRODUCTS_REQUEST",
-  GET_PRODUCTS_SUCCESS = "GET_PRODUCTS_SUCCESS",
-  GET_PRODUCTS_ERROR = "GET_PRODUCTS_ERROR",
-
-  POST_CART_PRODUCTS_REQUEST = "POST_CART_PRODUCTS_REQUEST",
-  POST_CART_PRODUCTS_SUCCESS = "POST_CART_PRODUCTS_SUCCESS",
-  POST_CART_PRODUCTS_ERROR = "POST_CART_PRODUCTS_ERROR",
-
-  GET_CART_PRODUCTS_REQUEST = "GET_CART_PRODUCTS_REQUEST",
-  GET_CART_PRODUCTS_SUCCESS = "GET_CART_PRODUCTS_SUCCESS",
-  GET_CART_PRODUCTS_ERROR = "GET_CART_PRODUCTS_ERROR",
-
-  DELETE_CART_PRODUCTS_REQUEST = "DELETE_CART_PRODUCTS_REQUEST",
-  DELETE_CART_PRODUCTS_SUCCESS = "DELETE_CART_PRODUCTS_SUCCESS",
-  DELETE_CART_PRODUCTS_ERROR = "DELETE_CART_PRODUCTS_ERROR",
-
-  GET_ORDERS_REQUEST = "GET_ORDERS_REQUEST",
-  GET_ORDERS_SUCCESS = "GET_ORDERS_SUCCESS",
-  GET_ORDERS_ERROR = "GET_ORDERS_ERROR",
-
-  GET_ORDER_REQUEST = "GET_ORDER_REQUEST",
-  GET_ORDER_SUCCESS = "GET_ORDER_SUCCESS",
-  GET_ORDER_ERROR = "GET_ORDER_ERROR",
-
-  POST_PAYMENT_PRODUCTS_REQUEST = "POST_PAYMENT_PRODUCTS_REQUEST",
-  POST_PAYMENT_PRODUCTS_SUCCESS = "POST_PAYMENT_PRODUCTS_SUCCESS",
-  POST_PAYMENT_PRODUCTS_ERROR = "POST_PAYMENT_PRODUCTS_ERROR",
-
-  POST_ORDER_DETAILS_REQUEST = "POST_ORDER_DETAILS_REQUEST",
-  POST_ORDER_DETAILS_SUCCESS = "POST_ORDER_DETAILS_SUCCESS",
-  POST_ORDER_DETAILS_ERROR = "POST_ORDER_DETAILS_ERROR",
-}
