+++ client/src/redux/cart/actionType.ts
@@ -1,0 +1,13 @@
+export enum CartActionType {
+  POST_CART_PRODUCTS_REQUEST = "POST_CART_PRODUCTS_REQUEST",
+  POST_CART_PRODUCTS_SUCCESS = "POST_CART_PRODUCTS_SUCCESS",
+  POST_CART_PRODUCTS_ERROR = "POST_CART_PRODUCTS_ERROR",
+
+  GET_CART_PRODUCTS_REQUEST = "GET_CART_PRODUCTS_REQUEST",
+  GET_CART_PRODUCTS_SUCCESS = "GET_CART_PRODUCTS_SUCCESS",
+  GET_CART_PRODUCTS_ERROR = "GET_CART_PRODUCTS_ERROR",
+
+  DELETE_CART_PRODUCTS_REQUEST = "DELETE_CART_PRODUCTS_REQUEST",
+  DELETE_CART_PRODUCTS_SUCCESS = "DELETE_CART_PRODUCTS_SUCCESS",
+  DELETE_CART_PRODUCTS_ERROR = "DELETE_CART_PRODUCTS_ERROR",
+}
