+++ client/src/redux/product/actionType.ts
@@ -1,0 +1,5 @@
+export enum ProductActionType {
+  GET_PRODUCTS_REQUEST = "GET_PRODUCTS_REQUEST",
+  GET_PRODUCTS_SUCCESS = "GET_PRODUCTS_SUCCESS",
+  GET_PRODUCTS_ERROR = "GET_PRODUCTS_ERROR",
+}
