+++ client/src/redux/store.ts
@@ -1,11 +1,15 @@
-import { createStore, applyMiddleware, compose } from 'redux';
-import thunk from 'redux-thunk';
-import reducers from './reducers';
+import { createStore, applyMiddleware, compose } from "redux";
+import thunk from "redux-thunk";
+import reducers from "./reducers";
 
 const composeEnhancers =
-  (process.env.NODE_ENV === 'development' &&
-    typeof window !== 'undefined' &&
+  (process.env.NODE_ENV === "development" &&
+    typeof window !== "undefined" &&
     (window as any).__REDUX_DEVTOOLS_EXTENSION_COMPOSE__) ||
   compose;
 
-export const store = createStore(reducers, {}, composeEnhancers(applyMiddleware(thunk)));
+export const store = createStore(
+  reducers,
+  {},
+  composeEnhancers(applyMiddleware(thunk))
+);
